pyproject.toml
[project]
name = "aim-challenge"
version = "0.1.0"
description = "Aim data-pipeline prototype — ingests heterogeneous sources, filters via hybrid RAG, generates per-Aim digests."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "openai>=1.55",
    "pinecone>=5.3",
    "feedparser>=6.0",
    "trafilatura>=2.0",
    "langchain-text-splitters>=0.3",
    "tenacity>=9.0",
    "python-dotenv>=1.0",
    "pydantic>=2.9",
    "google-cloud-firestore>=2.19",
    "google-cloud-bigquery>=3.27",
    "google-cloud-aiplatform>=1.80",
    "google-cloud-storage>=2.19",
    "numpy>=2.4.4",
]
